refactor: remove frame tabs component
