--- src/app/dashboard/create/page.tsx
@@ -1,16 +1,0 @@
-import { FrameTabs } from "./components/frame-tabs";
-
-export default function CreatePage() {
-  return (
-    <section className="container flex flex-col items-center">
-      <div className="text-center mb-12">
-        <h1 className="font-semibold text-3xl">Publish Your Frame</h1>
-        <p className="text-muted-foreground">
-          Let&apos;s supercharge your campaign!
-        </p>
-      </div>
-
-      <FrameTabs />
-    </section>
-  );
-}
