+++ src/app/dashboard/publish/components/frame-details.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { z } from "zod";
 import { toast } from "sonner";
 import { useForm } from "react-hook-form";
