+++ src/app/dashboard/publish/components/frame-preview.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import Image from "next/image";
 import { useRef, useState } from "react";
 import {
