+++ src/components/navbar.tsx
@@ -2,7 +2,7 @@ import Link from "next/link";
 
 export function Navbar() {
   return (
-    <nav className="py-12 flex justify-between items-center container">
+    <nav className="py-10 flex justify-between items-center container">
       <Link href="/" className="text-2xl font-black tracking-[-0.09em]">
         frame<span className="text-purple-600">yu</span>
       </Link>
